# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "data-loom.",
        "title": "data-loom API",
        "contact": {},
        "version": "v1.0"
    },
    "host": "localhost:8080",
    "basePath": "/data-loom/api/v1",
    "paths": {
...
        "/task/changelog": {
            "get": {
                "description": "查询任务状态变更历史极力",
                "tags": [
                    "策略"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "pageSize",
                        "name": "pageSize",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "任务ID",
                        "name": "task_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/services.TaskChangeLogService"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "common.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "返回码",
                    "type": "integer"
                },
                "data": {
                    "description": "返回数据"
                },
                "error": {
                    "description": "错误信息",
                    "type": "string"
                },
                "message": {
                    "description": "返回信息",
                    "type": "string"
                }
            }
        },
...
        "common.TaskStatusType": {
            "type": "string",
            "enum": [
                "scheduled",
                "supplement_failed",
                "waiting",
                "exec_check_failed",
                "executing",
                "success",
                "failed",
                "timeout"
            ],
            "x-enum-comments": {
                "TaskStatusExecCheckFailed": "执行前检查失败",
                "TaskStatusExecFailed": "执行失败",
                "TaskStatusExecuting": "执行中",
                "TaskStatusScheduled": "已排期",
                "TaskStatusSuccess": "执行成功",
                "TaskStatusSupplementFailed": "填充任务信息失败",
                "TaskStatusTimeout": "执行超时",
                "TaskStatusWaiting": "等待执行"
            },
            "x-enum-varnames": [
                "TaskStatusScheduled",
                "TaskStatusSupplementFailed",
                "TaskStatusWaiting",
                "TaskStatusExecCheckFailed",
                "TaskStatusExecuting",
                "TaskStatusSuccess",
                "TaskStatusExecFailed",
                "TaskStatusTimeout"
            ]
        },
...
        "services.TaskChangeLogService": {
            "type": "object",
            "properties": {
                "content": {
                    "description": "变化内容记录",
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "task_id": {
                    "description": "任务ID",
                    "type": "integer"
                },
                "task_status": {
                    "description": "任务状态",
                    "allOf": [
                        {
                            "$ref": "#/definitions/common.TaskStatusType"
                        }
                    ]
                },
                "time": {
                    "description": "创建时间",
                    "type": "string"
                },
                "user_name": {
                    "description": "创建人",
                    "type": "string"
                }
            }
        },
...
    }
}